### 2023-10-01 #5912 @editorialbot
[PRE REVIEW]: AeroAcoustics.jl: A julia package for aeroacoustics
**Submitting author:** @1oly (<a href="http://orcid.org/0000-0002-1596-107X">Oliver Lylloff</a>)
**Repository:** https://github.com/1oly/AeroAcoustics.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.3
