### 2023-10-03 #5912 @1oly
Thank you for starting the pre-review. I understand that many of you are busy, and appreciate your time and effort! I've looked at potential reviewers in your list, and the following looks like very capable candidates:
`@svchb` and
`@thejasvibr` (if willing to review julia code)

Potential reviewers, that have not signed up (yet), but would be really good:
`@ssfrr`, `@jfsantos`, `@nantonel`
If willing to review julia-code: `@fhchl`, `@gherold`.
