### 2023-10-10 #5912 @faroit
@kyleniemeyer i'm not an expert on Julia and also not very proficient in acoustics but I can take this once I have capacity (should be in the next 2 weeks)
