### 2024-01-26 #5912 @thejasvibr
Sounds very interesting. I'd be happy to be a reviewer for this package.

On Fri, Jan 26, 2024, 11:31 PM Fabian-Robert Stöter <
***@***.***> wrote:

> 👋 @thejasvibr <https://github.com/thejasvibr> @svchb
> <https://github.com/svchb> @ssfrr <https://github.com/ssfrr> @jfsantos
> <https://github.com/jfsantos> @nantonel <https://github.com/nantonel> -
> would any of you be willing to review this submission for JOSS? The JOSS
> review process takes place on GitHub and focuses on the software and a
> short paper. We carry out our checklist-driven reviews here in GitHub
> issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The software under review is a julia package
> https://github.com/mei-friend/mei-friend
> <https://github.com/openjournals/joss-reviews/issues/mei-friend>
>
> This issue is a "pre-review" issue in which reviewers are assigned. Once
> sufficient reviewers are recruited we will open a dedicated review issue
> where the review will take place.
>
