### 2024-01-27 #5912 @thejasvibr
Can confirm i'd be happy to review the Aeroacoustics package

On Sat, Jan 27, 2024, 11:25 AM Fabian-Robert Stöter <
***@***.***> wrote:

> @svchb <https://github.com/svchb> I linked the wrong submission. Sorry
> for that. Can you confirm that you could review
> https://github.com/1oly/AeroAcoustics.jl ? Thank you
>
