### 2024-01-28 #5912 @1oly
Hi @faroit. Thanks for handling this 👍

I can confirm there's no conflict of interest.

> Hi @1oly sorry again for the long waiting time 🙇
>
> Thank you already for providing a list of reviewers with domain knowledge. In addition to your proposal in [#5912 (comment)](https://github.com/openjournals/joss-reviews/issues/5912#issuecomment-1744527828) I would also like to ask the following julia audio devs:
>
> * jongwook
> * martinholters
> * gummif
>
> Can you please check all of them are ok wrt the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)?
