### 2024-02-07 #5912 @faroit
👋 @jongwook @martinholters @gummif - would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is the following julia package [https://github.com/1oly/AeroAcoustics.jl](https://github.com/1oly/AeroAcoustics.jl)
