### 2024-02-11 #5912 @faroit
@nantonel The review process will start in a separate issue when we have a second reviewer. I am going to assign one by next week. After that, the reviews are expected to be completed in about 2-4 weeks.

Thanks a lot for helping JOSS!
