### 2024-02-21 #6390 @editorialbot
[REVIEW]: AeroAcoustics.jl: A julia package for aeroacoustics
**Submitting author:** @1oly (<a href="http://orcid.org/0000-0002-1596-107X">Oliver Lylloff</a>)
**Repository:** https://github.com/1oly/AeroAcoustics.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.4
