### 2024-03-01 #6390 @faroit
👋 @nantonel @thejasvibr thanks again for helping out. Let me know if you need some help starting the review process.

If you have comments for the submission, I would generally appreciate if you add issues to https://github.com/1oly/AeroAcoustics.jl and add a reference [back here](https://github.com/openjournals/joss-reviews/issues/6390) so that this issue doesn't become too cluttered.
