### 2024-03-21 1oly/AeroAcoustics.jl#6 @thejasvibr
no example data in repo to run notebook
As I try to run Cell 2 of the example notebook, this error pops up.
Also, having looked into the repo as it stands - doesn't seem like the data files are in version control.

![image](https://github.com/1oly/AeroAcoustics.jl/assets/12443569/291bd7a9-c20d-4776-a324-0f1a61196fb1)

- related to openjournals/joss-reviews#6390
