### 2024-03-21 1oly/AeroAcoustics.jl#7 @thejasvibr
online docs on juliahub - up to date/duplicate?
Hi,
This is in relation to openjournals/joss-reviews#6390.

There seems to be [online documentation](https://docs.juliahub.com/AeroAcoustics/M8loF/0.2.3/) that's not mentioned in the Github repo's README. Found it through a quick online search, and other future users may potentially do so too - and thus the question.

Is this online docs page up to date - or was it a trial that was not completed. It seems like some parts of the README and example notebook are online currently. In my opinion, an online docs page is way more convincing than a user having to open up Jupyter Lab, wait for it all to open, and then open a notebook file. A nice online docs page is a reliable indicator into how user-friendly a package is, and can attract potential users to install the package and try it out themselves.

The decision to keep an online docs or not is up to you of course. Either way, it would be good to either take down the page if there's no plan to keep it up to date, or go all in and make it detailed enough to be useful.
