### 2024-03-21 1oly/AeroAcoustics.jl#8 @thejasvibr
instructions to run test missing
Hi,
This is in relation to https://github.com/openjournals/joss-reviews/issues/6390.

There is a ```tests``` folder with scripts to run test, however there aren't any instructions on how to setup the tests and run them. A separate Markdown file either in the home folder or in the ```tests``` folder would be great.
