### 2024-04-02 1oly/AeroAcoustics.jl#6 @1oly
Hi @thejasvibr ! Thanks for waiting. The data for the example is first downloaded from an external repo when you run the tests. This is not documented, I'm sorry about that. I will update the README and the example notebook to reflect this. If you want to test it out you should invoke the julia package manager by `]` (changing the prompt to `pkg>`) and write:
```
pkg> test AeroAcoustics
```
Then the data is downloaded and should subsequently be available for the example script...
This should also somewhat address your other issue about tests. I will working on a better documentation of this in the coming week.
