### 2024-04-04 1oly/AeroAcoustics.jl#6 @1oly
Hi again @thejasvibr, could I ask you to report back about the above suggestion, when time permits? If it works, I will add the instructions to the README and script.
