### 2024-04-08 1oly/AeroAcoustics.jl#6 @1oly
Hi @thejasvibr Thank you. I completely agree, and will find an alternative solution about the data. But for now, I'm hoping to get the current script running, since there might be other issues there to be found.

Did the tests complete?
If yes, then you were in fact able to download the file, but it's simply not found by the script.
