### 2024-04-08 1oly/AeroAcoustics.jl#6 @thejasvibr
The tests passed successfully, so I guess the data isn't being located
somehow as you suspect.

On Mon, Apr 8, 2024, 12:38 PM Oliver Lylloff ***@***.***>
wrote:

> Hi @thejasvibr <https://github.com/thejasvibr> Thank you. I completely
> agree, and will find an alternative solution about the data. But for now,
> I'm hoping to get the current script running, since there might be other
> issues there to be found.
>
> Did the tests complete?
> If yes, then you were in fact able to download the file, but it's simply
> not found by the script.
>
