### 2024-04-08 1oly/AeroAcoustics.jl#6 @nantonel
In the [doc](https://1oly.github.io/AeroAcoustics.jl/dev/) I'd change `csm_file = joinpath(dirname(@__DIR__),"test","data","test1_csm.h5")` to something like:
```
csm_file = joinpath(replace(dirname(pathof(AeroAcoustics)),"src"=>"test"),"data","test1_csm.h5")
```
then I could run the code snippets except for `sourceintegration(x,E,int_region)`.
