### 2024-04-09 1oly/AeroAcoustics.jl#9 @nantonel
[Docs] Source integration - undefined `x`
In this [doc section](https://1oly.github.io/AeroAcoustics.jl/dev/#Source-integration-1) `x` is not defined in the code snippet. If an example can be given I think it would improve the docs. Thanks

Related to review: https://github.com/openjournals/joss-reviews/issues/6390
