+++ README.md
@@ -44,6 +44,14 @@ First install [julia](http://julialang.org) and start julia in a terminal, [VS c
 pkg> add AeroAcoustics
 ```
 the package manager `pkg>` can be accessed by typing `]`.
+## Tests
+This package constains tests used for CI, but can also be used to check if the package is working properly when installed. To run the tests, after adding the package, activate the package manager, by typing `]`, and write
+```
+pkg> test AeroAcoustics
+```
+The test suite will download an external file, that is stored in `test/data`.
+## Documentation
+The [documentation](https://1oly.github.io/AeroAcoustics.jl/dev/) gives an introduction to the packages and the API.
 ## Contribution
 Contributions are welcome! Issues are tracked on [Github issue tracker](https://github.com/1oly/AeroAcoustics.jl/issues). If you want to add an new algorithm, you can fork this package and start developing your code and test it.
 
