### 2024-04-10 1oly/AeroAcoustics.jl#9 @1oly
Thanks for the comment !

This should be clarified with #10
It's currently on master, will tag it later. Feel free to open another issue, if it is still not clear.
