+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"reference/#API-Reference-1","page":"API Reference","title":"API Reference","text":"","category":"section"},{"location":"reference/#General-1","page":"API Reference","title":"General","text":"","category":"section"},{"location":"reference/#","page":"API Reference","title":"API Reference","text":"Modules = [AeroAcoustics]\nPages   = [\"AeroAcoustics.jl\",\"types.jl\",\"utils.jl\"]","category":"page"},{"location":"reference/#AeroAcoustics.DR!-Tuple{FreqArray}","page":"API Reference","title":"AeroAcoustics.DR!","text":"DR!(CSM::FreqArray)\n\nApply diagonal removal to CSM.\n\n\n\n\n\n","category":"method"},{"location":"reference/#AeroAcoustics.DR-Tuple{FreqArray}","page":"API Reference","title":"AeroAcoustics.DR","text":"DR(CSM::FreqArray)\n\nApply diagonal removal to CSM and new array.\n\n\n\n\n\n","category":"method"},{"location":"reference/#AeroAcoustics.narrow2oct-Tuple{FreqArray, Any}","page":"API Reference","title":"AeroAcoustics.narrow2oct","text":"narrow2oct(x::FreqArray,n,nomial::Bool=true;psd=false)\n\nSum narrow band spectrum to 1/n octave bands given narrow band frequencies f in x.\n\n\n\n\n\n","category":"method"},{"location":"reference/#AeroAcoustics.Environment","page":"API Reference","title":"AeroAcoustics.Environment","text":"Environment\n\nThe Environment struct is required for most methods and defines the geometrical setup, constants, and stores the relevant data together. The microphone array is assumed to be at the center of the coordinate system. The fields are\n\nArguments\n\nRequired:\n\nmicgeom::Array{T,2}: (x,y,z) coordinates for microhone array.\nz0::Real: distance to source plane from microphone array.\nCSM::FreqArray: Cross-spectral matrix, size M x M x Nf, where M is the number of microphones and Nf is the number of frequency bins.\n\nOptional:\n\nflim::Tuple=extrema(CSM.fc): Frequency limits (fmin,fmax).\nNx::Integer=21: Number of computational gridpoint in x direction.\nNy::Integer=21: Number of computational gridpoint in y direction.\nxlim::Tuple=(-1.,1.): Cartesian x-coordinate limits.\nylim::Tuple=(-1.,1.): Cartesian y-coordinate limits.\nwv=ones(size(micgeom,2)): Microphones on/off. \nwc::Bool=false: Microphones coherence weighting.\ndr::Bool=false: Diagonal removal of CSM\nw=ones(M,Nf): Microphone weights. If wc=true it calculates coherence weighting.\nshear::Bool = false: Amiet phase correction\nampcorr::Bool = shear: Amiet amplitude correction (only applies when shear = true)\nc::Real=343.: Speed of sound [m/s].\nMa::Real=0.0: Mach number (sign determines flow direction)\nh::Real=0.0: Distance from array center to shear layer (Amiet correction) should be supplied when Ma != 0.\nmulti_thread::Bool = false: Enable multi-threading via the ThreadsX package.\n\n\n\n\n\n","category":"type"},{"location":"reference/#AeroAcoustics.coherence_weights-Tuple{Any}","page":"API Reference","title":"AeroAcoustics.coherence_weights","text":"coherence_weights(CSM)\n\nCalculate microphone weights according to mean mututal-coherence.\n\nAmaral et al. “Slat Noise from an MD30P30N Airfoil at Extreme Angles of Attack,” AIAA Journal, vol. 56, no. 3, pp. 964–978,
[…]
