### 2024-04-10 1oly/AeroAcoustics.jl#8 @1oly
Hi, just pushed #10 with clarifications in the README on how to run tests. It's available on the master branch, will tag it later. Closing. Feel free to open the issue again if it is not clear.
