### 2024-04-10 1oly/AeroAcoustics.jl#7 @1oly
Hi, and thanks for the comment!
This should be a little clearer with #10

When the package is registered in the julia metadata, the docs are copied to juliahub. These are simply the same as the ones in this repo. It's not the most elaborate docs, but I intend to add some more examples.
