### 2024-04-10 1oly/AeroAcoustics.jl#6 @1oly
Hi both,
Thanks for the comments and suggested change in the code. I have pushed #10 with some updates to the docs and example script. Hope you can run it now.

Closing, but please reopen if the issue persist!
