### 2024-04-11 1oly/AeroAcoustics.jl#11 @nantonel
[Docs] add community guidelines
Could you please add some text addressing the following?

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Issue opened as part of review: https://github.com/openjournals/joss-reviews/issues/6390
