### 2024-04-11 1bdad80
update readme with community guidelines
