### 2024-04-11 1oly/AeroAcoustics.jl#11 @1oly
Hi @nantonel Thanks for the suggestion. Just pushed this to master:

## Community and contributions
Any question regarding the installation, use or extension of the code can be posted in the [Github issue tracker](https://github.com/1oly/AeroAcoustics.jl/issues).
If you want to add an new algorithm, you can fork this package and start developing your code and test it. If you wish to contribute to the general development, you can look for open issues or reach out to discuss further. Contributions are very welcome!
