+++ LICENSE.md
@@ -1,6 +1,6 @@
 The AeroAcoustics.jl package is licensed under the MIT "Expat" License:
 
-> Copyright (c) 2019-2022: ollyl.
+> Copyright (c) 2019-2024: ollyl.
 >
 > Permission is hereby granted, free of charge, to any person obtaining a copy
 > of this software and associated documentation files (the "Software"), to deal
