### 2024-05-16 #6390 @faroit
@1oly thanks for the long wait. The submission looks good now and we can finally move forward 👏 I just did check the paper again for grammar and spellings and created a pull request that you can merge in if you like.

@1oly when the paper is ready, if necessary then please make a new release of the main repo that includes all of the changes that have resulted from the review. Please report the version number here. This is only required if there is a change to the software. The paper branch doesn't need a new version.

After that, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the archive version, please make sure that:

- The authors of the archive are the same as the JOSS paper authors
