### 2024-05-17 #6390 @1oly
Hi @faroit Thanks!

Version is still 0.2.4.
Uploaded to Zenodo with doi: https://zenodo.org/doi/10.5281/zenodo.11207482
