### 2024-05-17 #6390 @faroit
The two missing DOIs were checked there doesn't seem to be a good replacement.
