### 2024-05-24 #6390 @kyleniemeyer
Hi @1oly, sorry for the delay here! I lost track of this one.

In both the paper title and in the body, please capitalize "Julia". I also noticed an instance of "Python" in the references that needs to be capitalized. Once you fix those little issues, let me know here and I will accept this.
