### 2024-05-25 c96a2d3
fix capitalization of Julia and Python
