+++ paper/paper.bib
@@ -53,7 +53,7 @@ doi       = "10.1137/141000671"
    month = {1},
    pages = {50-58},
    publisher = {Elsevier Ltd},
-   title = {A Python framework for microphone array data processing},
+   title = {A {Python} framework for microphone array data processing},
    volume = {116},
    year = {2017},
 }
