### 2024-05-25 342c027
fix capitalization of Julia and Python
