+++ paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 # Summary
 Aeroacoustics is an interdisciplinary research field situated between aerodynamics and acoustics. It focuses on understanding the sound generated by the motion of air around solid bodies and plays an important role in research areas such as automotive, aerospace, and wind energy sciences [@Allen2002c]. However, the airflow responsible for sound generation typically generates unwanted background noise, which masks the true acoustic emission of the test item from measurement equipment. Therefore, isolating the sound emitted by the test item becomes essential to obtain reliable evaluations, which is pivotal in the development of noise-reducing technologies for cars, airplanes, and wind turbines. One common solution strategy involves the use of *acoustic images* computed using microphone arrays. The acoustic images help visualize and quantify the spatial distribution of sound across the test item, and enables precise assessment of acoustic emissions without contamination from background noise [@Glegg2017].
 
-[`AeroAcoustics.jl`](https://github.com/1oly/AeroAcoustics.jl) is a julia [@Bezanson2017] package for working with microphone arrays and acoustic imaging techniques. Methods for computing acoustic images often share the same codebase and utility functions, making them suitable candidates for a modular software package. It is the intention, that `AeroAcoustics.jl` can serve as a foundation for benchmarking existing methods and developing new ones within the aeroacoustics community. In addition to that, `AeroAcoustics.jl` is also a good candidate for teaching and entry-level programmers due to its simple API.
+[`AeroAcoustics.jl`](https://github.com/1oly/AeroAcoustics.jl) is a Julia [@Bezanson2017] package for working with microphone arrays and acoustic imaging techniques. Methods for computing acoustic images often share the same codebase and utility functions, making them suitable candidates for a modular software package. It is the intention, that `AeroAcoustics.jl` can serve as a foundation for benchmarking existing methods and developing new ones within the aeroacoustics community. In addition to that, `AeroAcoustics.jl` is also a good candidate for teaching and entry-level programmers due to its simple API.
 
 # Statement of need
 The `AeroAcoustics.jl` package provides a collection of commonly used acoustic imaging techniques, developed by the aeroacoustic community, such as conventional beamforming [@Johnson1993a], DAMAS [@Brooks2006], and CleanSC [@Sijtsma2007]. The `AeroAcoustics.jl` package also includes newer techniques found in the literature, e.g., fista [@Lylloff2015] and mean mutual-coherence weights [@Amaral2018]. 
