### 2024-05-29 66987fb
update with joss citation info
