### 2024-05-29 a09653c
CompatHelper: bump compat for LazyArrays to 2, (keep existing compat)
