+++ Project.toml
@@ -21,7 +21,7 @@ ThreadsX = "ac1d9e8a-700a-412c-b207-f0111f4b6c0d"
 DSP = "0.6.7, 0.7"
 Distances = "0.10"
 FFTW = "1.1"
-LazyArrays = "0.16, 0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1"
+LazyArrays = "0.16, 0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1, 2"
 NLsolve = "4.4"
 Parameters = "0.12.3"
 ThreadsX = "0.1.10"
