+++ README.md
@@ -72,8 +72,8 @@ Another noteworthy library for microphone array measurements is [Acoular](http:/
 If you use this package in your work, please cite the following:
 ```  
 @article{lylloff2024aeroacoustics, 
-doi = {10.21105.joss.06390}, 
-url = {https://doi.org/10.21105/joss.06390}, 
+doi = {10.21105/joss.06390}, 
+url = {https://joss.theoj.org/papers/10.21105/joss.06390}, 
 year = {2024}, 
 publisher = {The Open Journal}, 
 journal = {Journal of Open Source Software},
