### 2024-05-29 #6390 @kyleniemeyer
@1oly I think there may be some CrossRef issues at the moment. Nothing for you to do - the paper is officially published: https://joss.theoj.org/papers/10.21105/joss.06390
